An other go at the 'wide character' bug
It turns out, that in the "old" application/x-www-form-urlencoded
interface, the bytea fields are not properly encoded and need an extra
round of encoding, before  they enter the database. The RESTish/JSONRPC2
interface doesn't seem to have this problem, so we take care to only do
this for the form-post interface.